# api/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Error:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  domain.Image:
    properties:
      data:
        items:
          type: integer
        type: array
      hash:
        type: string
      id:
        type: string
    type: object
...
  dto.Image:
    properties:
      image:
        items:
          type: integer
        type: array
    required:
    - image
    type: object
...
host: aboba.com
info:
  contact: {}
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger CRUD Home appliance store API
  version: "1.0"
paths:
...
  /api/v1/images:
    get:
      consumes:
      - application/json
      description: The endpoint for retrieve all registered images in system
      parameters:
      - description: limit get images
        in: query
        name: limit
        required: true
        type: integer
      - description: offset get images
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get all images
      tags:
      - images
    post:
      consumes:
      - application/json
      description: 'Image created from JSON or XML, for create endpoint required:
        image'
      parameters:
      - description: Image data
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/domain.Image'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Create image
      tags:
      - images
  /api/v1/images/{id}:
    delete:
      consumes:
      - application/json
      description: The endpoint for deleting image data by ID
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Delete image
      tags:
      - images
    get:
      consumes:
      - application/json
      description: The endpoint for retrieve registered image in system by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get all images
      tags:
      - images
    patch:
      consumes:
      - application/json
      description: The endpoint for updating image data by ID to a new image given
        by the user
      parameters:
      - description: New image data
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/domain.Image'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Update image
      tags:
      - images
...
swagger: "2.0"